PB-28 add IStorage, IUserStorage and UserStorage handlers
